fix update hunt, questions and other files

+++ backend/som/api/survey/update_survey.py
@@ -30,6 +30,8 @@ def save_files(files: list[UploadFile], file_type: str, survey_number: int):
     os.makedirs(save_dir, exist_ok=True)
 
     for file in files:
+        if not file:
+            continue
         file_path = save_dir / file.filename
         with open(file_path, "wb") as buffer:
             shutil.copyfileobj(file.file, buffer)
@@ -40,6 +42,8 @@ def delete_files(files: list[UploadFile], file_type: str, survey_number: int):
         return
     
     for file in files:
+        if not file:
+            continue
         file_path = UPLOAD_DIR / str(survey_number) / file_type / file.filename
         if file_path.exists():
             file_path.unlink()
@@ -79,8 +83,8 @@ def update_survey(
     create_survey_files(survey_number_id=survey_number.id, file_extension="other", files=other_files, db=db)
 
     try:
-        save_files(questions_file, "questions", survey_number=survey_number.number)
-        save_files(hunt_file, "hunt", survey_number=survey_number.number)
+        save_files([questions_file], "questions", survey_number=survey_number.number)
+        save_files([hunt_file], "hunt", survey_number=survey_number.number)
         save_files(csv_files, "csv", survey_number=survey_number.number)
         save_files(ecg_files, "ecg", survey_number=survey_number.number)
         save_files(hr_files, "hr", survey_number=survey_number.number)
